dev/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
cloudinary==1.30.0
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
Django==4.1.5
django-appconf==1.0.5
django-cloudinary-storage==0.3.0
django-colorfield==0.8.0
django-nextjs==2.2.1
django-tinymce==3.5.0
djangorestframework==3.14.0
frozenlist==1.3.3
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jsmin==3.0.1
lxml==4.9.2
multidict==6.0.4
Pillow==9.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
python-dotenv==0.21.1
pytz==2022.7
rcssmin==1.1.1
requests==2.28.1
rjsmin==1.2.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
waitress==2.1.2
websockets==10.4
yarl==1.8.2
zope.interface==5.5.2
